Rewritten some parts of the documentation
The README, TODO and Developer Guidelines documents were rewritten to
better match the actual progress status of the project. Specifically,
all installation instructions were removed from README, as they are
not working right now. The TODO list was written from scratch.
The .pdf output from the user manual was removed from the repository
and added to the corresponding .gitignore.