Adding a randomized search feature to KD-Trees
We go downwards searching for a vector stochastically close
to the input vector. For each vector in a middle node we do
a random test, biased on the vectors' distance. If it passes
we return that vector, otherwise, we go left or right also based
on a random test biased of the input vector's distance to the
cut plane. We expect this test to return the closest vectors
with high probability with only a few tests.